+++ include/menu.h
@@ -10,7 +10,7 @@ public:
   virtual void Draw() = 0; // Draw the menu on the screen
   u8 GetCurrentPage() const;
   u8 GetPageCount() const;
-  virtual void HandleInput(u16 keys) = 0; // Handle input events
+  virtual void HandleInput(u32 keys) = 0; // Handle input events
   void SelectItem(u8 index);
 
   class App *app; //! Pointer to the application instance.
